Hide scores on scoresheet, renamed component files to match current terminology
