+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cite-ui",
-  "version": "1.3.7-rc1",
+  "version": "1.3.7-rc2",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
