--- src/app/components/score-card/score-card.component.spec.ts
@@ -1,29 +1,0 @@
-// Copyright 2022 Carnegie Mellon University. All Rights Reserved.
-// Released under a MIT (SEI)-style license, please see LICENSE.md in the
-// project root for license information or contact permission@sei.cmu.edu for full terms.
-
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { ScoreCardComponent } from './score-card.component';
-
-describe('ScoreCardComponent', () => {
-  let component: ScoreCardComponent;
-  let fixture: ComponentFixture<ScoreCardComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ScoreCardComponent ]
-    })
-      .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ScoreCardComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-
