+++ src/app/components/score-summary/score-summary.component.scss
@@ -1,0 +1,81 @@
+// Copyright 2022 Carnegie Mellon University. All Rights Reserved.
+// Released under a MIT (SEI)-style license, please see LICENSE.md in the
+// project root for license information or contact permission@sei.cmu.edu for full terms.
+
+.score-card-container {
+  display: flex;
+  flex-flow: column;
+  justify-content: flex-start;
+  align-self: flex-start;
+  text-align: center;
+  padding: 10px;
+  height: 100%;
+}
+
+.sticky-card {
+  position: -webkit-sticky;
+  position: sticky;
+  top: 0;
+}
+
+.title-font {
+  font-size: 20px;
+  font-weight: bold;
+  justify-self: center;
+}
+
+.cite-icon {
+  transform: scale(0.8);
+  width: 60px;
+  margin-bottom: 10px;
+}
+/*
+.mat-table {
+  border: 1px double black;
+}
+*/
+.black {
+  background-color: black;
+  color: white;
+  height: 80px;
+  text-align: left;
+}
+.red {
+  background-color: darkred;
+  color: white;
+  height: 80px;
+  text-align: left;
+}
+.orange {
+  background-color: darkorange;
+  color: black;
+  height: 80px;
+  text-align: left;
+}
+.yellow {
+  background-color: yellow;
+  color: black;
+  height: 80px;
+  text-align: left;
+}
+.green {
+  background-color: lightgreen;
+  color: black;
+  height: 80px;
+  text-align: left;
+}
+.white {
+  background-color: white;
+  color: black;
+  min-height:80px;
+  text-align: left;
+}
+
+.highlight {
+  background-color: dimgrey;
+}
+
+.mat-row {
+  height: 100% !important;
+  align-items: stretch;
+}
