+++ src/app/components/score-summary/score-summary.component.spec.ts
@@ -1,0 +1,29 @@
+// Copyright 2022 Carnegie Mellon University. All Rights Reserved.
+// Released under a MIT (SEI)-style license, please see LICENSE.md in the
+// project root for license information or contact permission@sei.cmu.edu for full terms.
+
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { ScoreSummaryComponent } from './score-summary.component';
+
+describe('ScoreSummaryComponent', () => {
+  let component: ScoreSummaryComponent;
+  let fixture: ComponentFixture<ScoreSummaryComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ ScoreSummaryComponent ]
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ScoreSummaryComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
+
