+++ src/app/components/scoresheet/scoresheet.component.spec.ts
@@ -1,0 +1,29 @@
+// Copyright 2022 Carnegie Mellon University. All Rights Reserved.
+// Released under a MIT (SEI)-style license, please see LICENSE.md in the
+// project root for license information or contact permission@sei.cmu.edu for full terms.
+
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { ScoresheetComponent } from './scoresheet.component';
+
+describe('ScoresheetComponent', () => {
+  let component: ScoresheetComponent;
+  let fixture: ComponentFixture<ScoresheetComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ ScoresheetComponent ]
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ScoresheetComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
+
