--- src/app/components/scoring-model/scoring-model.component.spec.ts
@@ -1,29 +1,0 @@
-// Copyright 2022 Carnegie Mellon University. All Rights Reserved.
-// Released under a MIT (SEI)-style license, please see LICENSE.md in the
-// project root for license information or contact permission@sei.cmu.edu for full terms.
-
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { ScoringModelComponent } from './scoring-model.component';
-
-describe('ScoringModelComponent', () => {
-  let component: ScoringModelComponent;
-  let fixture: ComponentFixture<ScoringModelComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ScoringModelComponent ]
-    })
-      .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ScoringModelComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
-
